add an output masking feature to grepalign
Masking might be useful when caring only about the content before or
after the match while some of the lines are so long that they require
a terminal window resize or a pipe to less -S.  It might also be useful
as a quick and a rather specific cutting tool.

Also, refactored a few variable names.